Simplify code and add tests
