+++ WordPress/src/main/java/org/wordpress/android/ui/DeeplinkWebViewActivity.kt
@@ -49,4 +49,3 @@ class DeeplinkWebViewActivity : WebViewActivity() {
         }
     }
 }
-
